Improved setting of update status.
Formerly the update status in plugin_onlineshop_productindex_store_elastic was only
updated when ElasticSearchClient::bulk() would return an error for the whole batch.
This results in continuing to exist error messages/status codes for a whole batch of
datasets from older bulk indexing operations even if the error ceases to exist.

Now the status get's updated for each dataset with it's statuscode and null when
the error field is unset, or the json-encoded content of "error" when it's set.