ElasticSearch ProductList: Fixed resetting products
When calling getItems(), the already fetched products are resetted unnecessarily. This is due to setLimit() and setOffset() resetted it despite of neither limit nor offset have been changed.

This can result in a superfluous query to ES when simply rendering a pagination control.

This fix tries to be consistent with the DefaultMysql implementation: Reset the products only if offset/limit has been changed.