Bugfix(PricingRules): Prevent Error when adding a new Pricing Rule with a name that exists
When you create a Pricing Rule with an name that already exists the backend interface is not loading the pricing rules anymore because it inserts a row with "null" values.

The reason is because the create() call is successfully but at the update() call you get an errror because a pricing Rule with that name already exists. Solution -> wrap it into a transaction